MAINT: Collect model types in models.go
